remove decimal kas from totalsupply api

+++ http.go
@@ -42,7 +42,7 @@ func HttpServe() {
 			"</pre></body></html>"
 		} else {
 			html = "<html><head><meta name=\"color-scheme\" content=\"light dark\"></head><body><pre style=\"word-wrap: break-word; white-space: pre-wrap;\">" +
-			fmt.Sprintf("%.8f", float64(supply.CirculatingSompi)/constants.SompiPerKaspa) +
+			fmt.Sprintf("%d", uint64(supply.CirculatingSompi)/constants.SompiPerKaspa) +
 			"</pre></body></html>"
 		}
 		w.Write([]byte(html))
